+++ source/main/index.js
@@ -11,7 +11,6 @@ import {
   generateWalletMigrationReport,
 } from './utils/setupLogging';
 import { handleDiskSpace } from './utils/handleDiskSpace';
-// import { handleHardwareWalletDevices } from './ipc/getHardwareWalletChannel';
 import { handleCheckBlockReplayProgress } from './utils/handleCheckBlockReplayProgress';
 import { createMainWindow } from './windows/main';
 import { installChromeExtensions } from './utils/installChromeExtensions';
@@ -175,16 +174,6 @@ const onAppReady = async () => {
     isReady: hwHandler.isReady,
     hwHandler,
   });
-
-  // @TODO - uncomment once ledger should be enabled
-  // Ledger INIT
-  // console.debug('[HW-DEBUG] IS ENABLED: ', { isHardwareWalletSupportEnabled, isLedgerEnabled });
-  // if (isHardwareWalletSupportEnabled, isLedgerEnabled) {
-  //   const handleCheckHardwareWalletDevices = handleHardwareWalletDevices(
-  //     mainWindow
-  //   );
-  //   await handleCheckHardwareWalletDevices();
-  // }
 
   await handleCheckBlockReplayProgress(mainWindow, launcherConfig.logsPrefix);