+++ source/renderer/app/stores/HardwareWalletsStore.js
@@ -820,6 +820,7 @@ export default class HardwareWalletsStore extends Store {
         // Special case. E.g. device unplugged before cardano app is opened
         // Stop poller and re-initiate connecting state / don't kill devices listener
         this.stopCardanoAdaAppFetchPoller();
+
         runInAction(
           'HardwareWalletsStore:: Re-run initiated connection',
           () => {
@@ -830,6 +831,7 @@ export default class HardwareWalletsStore extends Store {
       } else if (error.code === 'DEVICE_PATH_CHANGED' && error.path) {
         // Special case on Windows where device path changes after opening Cardano app
         // Stop poller and re-initiate connecting state / don't kill devices listener
+        logger.debug('[HW-DEBUG] Update LC data with new path');
         this.stopCardanoAdaAppFetchPoller();
 
         const pairedDevice = find(
@@ -837,6 +839,7 @@ export default class HardwareWalletsStore extends Store {
           (recognizedDevice) => recognizedDevice.path === path
         );
 
+        // Update device with new path - LC
         await this._setHardwareWalletDevice({
           deviceId: pairedDevice.id,
           data: {
@@ -845,6 +848,44 @@ export default class HardwareWalletsStore extends Store {
             isPending: false,
           },
         });
+
+        // Update connected wallet data with new path - LC
+        if (walletId) {
+          logger.debug('[HW-DEBUG] Update connected wallet data with new path');
+          const hardwareWalletConnectionData = get(
+            this.hardwareWalletsConnectionData,
+            walletId
+          );
+
+          if (hardwareWalletConnectionData) {
+            logger.debug(
+              '[HW-DEBUG] Update connected wallet data with new path - Set to LC'
+            );
+            await this._setHardwareWalletLocalData({
+              walletId,
+              data: {
+                ...hardwareWalletConnectionData,
+                path: error.path,
+                device: {
+                  ...hardwareWalletConnectionData.device,
+                  path: error.path,
+                },
+              },
+            });
+          }
+        }
+
+        if (this.isTransactionInitiated) {
+          logger.debug(
+            '[HW-DEBUG] Update connected wallet data with new path - Set to LC'
+          );
+          runInAction(
+            'HardwareWalletsStore:: Change active device path for Transaction send',
+            () => {
+              this.activeDevicePath = error.path;
+            }
+          );
+        }
 
         this.cardanoAdaAppPollingInterval = setInterval(
           (devicePath, txWalletId) =>
@@ -1415,7 +1456,8 @@ export default class HardwareWalletsStore extends Store {
       });
     } else {
       logger.debug(
-        '[HW-DEBUG] HWStore - getCardanoAdaApp - from  initiateTransaction'
+        '[HW-DEBUG] HWStore - getCardanoAdaApp - from  initiateTransaction',
+        { devicePath }
       );
       if (walletId) {
         this.stopCardanoAdaAppFetchPoller();